requirements.txt
fastapi[standard]==0.112.0
python-multipart==0.0.9
pyjwt==2.9.0
python-dotenv
cryptography